% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addQuantileLevel.R
\name{addQuantileLevel}
\alias{addQuantileLevel}
\title{Add quantile level to data frame of deletion similarity results. Expects the input data frame to have a "Pvalue.FDR" column}
\usage{
addQuantileLevel(results = data.frame())
}
\arguments{
\item{results}{Correlation or enrichment test P-values}
}
\value{
results object with quantile level added as data frame column
}
\description{
Add quantile level to data frame of deletion similarity results. Expects the input data frame to have a "Pvalue.FDR" column
}
