% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLogical.R
\name{checkLogical}
\alias{checkLogical}
\title{Function to check validity of logical parameters}
\usage{
checkLogical(parName, parValue)
}
\arguments{
\item{parName}{Parameter name}

\item{parValue}{Input parameter value}
}
\description{
Function to check validity of logical parameters
}
