% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNumeric.R
\name{checkNumeric}
\alias{checkNumeric}
\title{Function to check validity of numeric parameters}
\usage{
checkNumeric(parName = "", parValue = NULL, minValue = NULL, maxValue = NULL)
}
\arguments{
\item{parName}{Parameter name}

\item{parValue}{Input parameter value}

\item{minValue}{Minimum valid value for parameter}

\item{maxValue}{Maximum valid value for parameter}
}
\description{
Function to check validity of numeric parameters
}
