% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDirectionalMatches.R
\name{computeDirectionalMatches}
\alias{computeDirectionalMatches}
\title{Used in hypergeometric analysis to identify genes that changed expression in the same direction}
\usage{
computeDirectionalMatches(signature1 = NA, signature2 = NA)
}
\arguments{
\item{signature1}{Strain 1 signature}

\item{signature2}{Strain 2 signature}
}
\value{
Data frame with information about overlapping genes in the two signatures
}
\description{
Used in hypergeometric analysis to identify genes that changed expression in the same direction
}
