% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegionEnrichment.R
\name{genomicRegionEnrichment}
\alias{genomicRegionEnrichment}
\title{Perform genomic region enrichment tests}
\usage{
genomicRegionEnrichment(
  genePositions = NULL,
  systematicNamesBG,
  systematicNames,
  includeMito = F,
  relativeTo = "telomere",
  rangeInKB = 25
)
}
\arguments{
\item{genePositions}{Table of genes and genomic positions (obtained using getGenePositions() if not specified)}

\item{systematicNamesBG}{Background set of systematic gene names to use for enrichment tests}

\item{systematicNames}{Set of systematic gene names to test for regional enrichment}

\item{includeMito}{Whether to include mitochondrial genes in the analysis}

\item{relativeTo}{Genomic region that will be tested for enrichment of input genes. Valid values are "telomere" or "centromere".}

\item{rangeInKB}{Window (in kilobases) for defining a gene as located in the region}
}
\description{
Perform genomic region enrichment tests
}
