% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllStrainNames.R
\name{getAllStrainNames}
\alias{getAllStrainNames}
\title{Get names of all deletion strains in the Deleteome}
\usage{
getAllStrainNames()
}
\value{
A character vector of Deleteome strain names
}
\description{
Get names of all deletion strains in the Deleteome
}
\examples{
# getAllStrainNames()

}
