% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenePositions.R
\name{getGenePositions}
\alias{getGenePositions}
\title{Get genomic locations for genes}
\usage{
getGenePositions(includeMito = F)
}
\arguments{
\item{includeMito}{Whether to include mitochondrial genes in the output}
}
\value{
Data frame with distances of genes (in bases) from closest telomere and from centromere
}
\description{
Get genomic locations for genes
}
