% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimilarStrainsByReciprocalCorrelation.R
\name{getSimilarStrainsByReciprocalCorrelation}
\alias{getSimilarStrainsByReciprocalCorrelation}
\title{Find deletion strains with similar expression profiles to a query strain using reciprocal correlation}
\usage{
getSimilarStrainsByReciprocalCorrelation(
  strain = "",
  outputDir = NA,
  minAbsLog2FC = 0,
  pDEGs = 0.05,
  pCor = 0.05,
  quantileCutoff = 0.1,
  returnTestValues = F,
  showMessages = F
)
}
\arguments{
\item{strain}{Name of Deleteome strain to analyze (query strain/query gene)}

\item{outputDir}{Directory in which to save similarity analysis results}

\item{minAbsLog2FC}{Log2 fold-change cutoff used to classify genes as differentially-expressed (the absolute value of log2 fold-change must be higher than minAbsLog2FC)}

\item{pDEGs}{P-value cutoff used to identify differentially-expressed genes}

\item{pCor}{P-value cutoff used to identify statistically significant correlation tests}

\item{quantileCutoff}{Quantile cutoff for selecting the Deleteome strain matches with highest confidence}

\item{returnTestValues}{If true, a data frame containing the correlation test results against each similar Deleteome strain is returned. If false, only the names of the similar strains are returned.}

\item{showMessages}{Whether to output progress messages}
}
\description{
Find deletion strains with similar expression profiles to a query strain using reciprocal correlation
}
\examples{
# getSimilarStrainsByReciprocalCorrelation(strain = "nup170",
#                                          outputDir = "...output directory path...",
#                                          minAbsLog2FC = 0,
#                                          pDEGs = 0.05,
#                                          pCor = 0.05,
#                                          quantileCutoff = 0.1)

}
