% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStrainNameORFmap.R
\name{getStrainNameORFmap}
\alias{getStrainNameORFmap}
\title{Getting mappings between Deleteome strain names and their open reading frames (ORFs)}
\usage{
getStrainNameORFmap()
}
\value{
A data frame with three columns (Strain, ORF, MappingType)
}
\description{
Getting mappings between Deleteome strain names and their open reading frames (ORFs)
}
