% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStrainSignature.R
\name{getStrainSignature}
\alias{getStrainSignature}
\title{Get data for a deletion strain's differentially-expressed genes}
\usage{
getStrainSignature(
  strain = "",
  minAbsLog2FC = 0,
  pDEGs = 0.05,
  consoleMessages = T
)
}
\arguments{
\item{strain}{Name of deletion strain as one-element character vector}

\item{minAbsLog2FC}{The absolute log2 fold-change threshold when selecting differentially-expressed genes (use 0 to include all genes)}

\item{pDEGs}{The p-value threshold to use when selecting differentially-expressed genes (use 1 to include all genes)}

\item{consoleMessages}{Whether to output console messages}
}
\value{
A data frame with systematic gene ids, gene names, M values (log2 fold-changes vs. WT), and p-values for a given deletion strain's transcriptome. Only microarrayed genes that meet M value and p-value cutoffs are included.
}
\description{
Get data for a deletion strain's differentially-expressed genes
}
\examples{
# getStrainSignature(strain = "nup170", minAbsLog2FC = 0, pDEGs = 0.05)

}
