% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmapDeleteomeMatches.R
\name{makeHeatmapDeleteomeMatches}
\alias{makeHeatmapDeleteomeMatches}
\title{Make a heatmap showing expression data for a query strain's signature compared to other Deleteome strains}
\usage{
makeHeatmapDeleteomeMatches(
  strain = NA,
  strainSignature = NA,
  otherStrains = NA,
  filePrefix = NA,
  titleDesc = "",
  minAbsLog2FCforTitle = NA,
  pDEGsForTitle = NA,
  pMatchesForTitle = NA,
  quantileForTitle = NA,
  subteloGenesOnly = F,
  clusterColumns = T,
  colFontSize = 1,
  showRowLabels = F,
  rowFontSize = 0.275,
  imageWidth = 3000,
  imageHeight = 2400,
  outputDir = NA,
  printToFile = T
)
}
\arguments{
\item{strain}{A Deleteome strain whose expression signature will be shown. Only used for plot titles and file names.}

\item{strainSignature}{The transcriptional signature of the deletion strain}

\item{otherStrains}{List of additional Deleteome strains to include in heatmap}

\item{filePrefix}{String that can be added to beginning of heatmap file name. Appears after strain name.}

\item{titleDesc}{Rationale for inclusion of additional strains in heatmap. Used in title of heatmap.}

\item{minAbsLog2FCforTitle}{Absolute log2 fold-change threshold used to get the mutant strain's signature. Only used in heatmap title and file name.}

\item{pDEGsForTitle}{P-value threhshold used to get the mutant strain's signature (if applicable). Only used in heatmap title and file name.}

\item{pMatchesForTitle}{P-value threshold used to select similar mutant strains (if applicable)}

\item{quantileForTitle}{Quantile threshold used to select similar strains (if applicable). Only used in heatmap title and file name.}

\item{subteloGenesOnly}{Whether to only include subtelomeric genes in the rows of the heatmap}

\item{clusterColumns}{Whether to automatically cluster the columns of the heatmap}

\item{colFontSize}{Font size for column labels}

\item{showRowLabels}{Whether to show row labels}

\item{rowFontSize}{Font size for row labels}

\item{imageWidth}{If writing to file, the Width of the saved image in pixels}

\item{imageHeight}{If writing to file, the Width of the saved image in pixels}

\item{outputDir}{If writing to file, output directory to save image}

\item{printToFile}{Whether to write heatmap to a file or show in new window}
}
\value{
Data matrix containing log2 fold-change expression data for heatmap
}
\description{
Make a heatmap showing expression data for a query strain's signature compared to other Deleteome strains
}
