% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGenomicPositionHistogram.R
\name{makeMountainLakePlot}
\alias{makeMountainLakePlot}
\title{Make a mountain lake plot: a histogram showing position of differentially-expressed genes relative to nearest telomere.}
\usage{
makeMountainLakePlot(
  strain = NULL,
  genePositions = NULL,
  relativeTo = "telomere",
  rangeInKB = 25,
  minAbsLog2FC = 0,
  pDEGs = 0.05,
  xMax = NA,
  yMax = 50,
  upColor = "#d53e4f",
  downColor = "#3288bd",
  showUpDownLabels = T,
  outputDir = NA,
  printToFile = T
)
}
\arguments{
\item{strain}{Name of Deleteome mutant strain to analyze}

\item{genePositions}{Table of genes and genomic positions (obtained using getGenePositions())}

\item{relativeTo}{Valid options are "telomere" or "centromere". If "telomere" specified, distance to closest telomere (right or left) is plotted.}

\item{rangeInKB}{Window (in kilobases) for defining a gene as either in the telomeric or centromeric region}

\item{minAbsLog2FC}{Log2 fold-change threshold for defining differentially-expressed genes}

\item{pDEGs}{P-value threshold for defining differentially-expressed genes}

\item{xMax}{X-axis maximum (in kilobases). If NA, uses maximum distance to closest telomere or to the centromere among all genes.}

\item{yMax}{Y-axis maximum}

\item{upColor}{Color to use for upregulated genes}

\item{downColor}{Color to use for downregulated genes}

\item{showUpDownLabels}{Whether to inclue "upregulated" and "downregulated" labels in plot}

\item{outputDir}{If writing to file, output directory to save image}

\item{printToFile}{Whether to write heatmap to a file or show in new window.}
}
\description{
Make a mountain lake plot: a histogram showing position of differentially-expressed genes relative to nearest telomere.
}
