% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegionEnrichment.R
\name{subtelomericEnrichment}
\alias{subtelomericEnrichment}
\title{Convenience function to test for enrichment of differentially-expressed genes in the subtelomeric region}
\usage{
subtelomericEnrichment(
  genePositions = NULL,
  systematicNamesBG,
  systematicNames,
  includeMito = F,
  rangeInKB = 25
)
}
\arguments{
\item{genePositions}{Table of genes and genomic positions (obtained using getGenePositions() if not specified)}

\item{systematicNamesBG}{Background set of systematic gene names to use for enrichment tests}

\item{systematicNames}{Set of systematic gene names to test for subtelomeric enrichment}

\item{includeMito}{Whether to include mitochondrial genes in the analysis}

\item{rangeInKB}{Window (in kilobases) for defining a gene as located in the subtelomeric region}
}
\description{
Convenience function to test for enrichment of differentially-expressed genes in the subtelomeric region
}
