% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOenrichmentOnStrains.R
\name{GOenrichmentOnStrains}
\alias{GOenrichmentOnStrains}
\title{Perform GO enrichment on a list of Deleteome strain names. By default, uses the list of strains in the Deleteome as the background for the enrichment tests. Performs enrichment test over GO:BP, GO:MF and GO:CC sub-ontologies.}
\usage{
GOenrichmentOnStrains(
  strains = c(),
  padjThresh = 0.1,
  useDeleteomeBackground = T
)
}
\arguments{
\item{strains}{A character vector of Deleteome strain names}

\item{padjThresh}{FDR-adjusted P-value cutoff for GO enrichment significance}

\item{useDeleteomeBackground}{gene names, not systematic names}
}
\description{
Perform GO enrichment on a list of Deleteome strain names. By default, uses the list of strains in the Deleteome as the background for the enrichment tests. Performs enrichment test over GO:BP, GO:MF and GO:CC sub-ontologies.
}
\examples{
# GOenrichmentOnStrains(strains = c("ctf18", "ctf8", "dcc1"), padjThresh = 0.1)

}
